User Interface Concept v1
Created a rough User Interface concept for the application. Includes a
control box for the basic controls and then an advance control box for
more advance controls (slides in from the left). Added the libraries for
this concept as well as a sample webgl cube for space filler.

Libraries are from http://srobbin.com/jquery-plugins/pageslide/