Fixed Issues 1 and 2 (partially)
Added cancel button and close button to LoadRemote Dialog window. The
cancel button brings it back to the WelcomeDialog, and the close button
goes to the ControlDialog. Changed the size of the Settings Dialog not
to take up the entire screen (was never meant to, just forgot to change
it when I commited). I will look into making it window that slide in
from a side.