Added Sample Audio File
Added sample audio file so that we don't need to load a file to test
(website needs to be hosted on a web server to work otherwise you get
the cross domain error). Added a Loading dialog box that shows the
current file is being loaded. Updated so the Welcome Dialog box
disappears after loading a file.