Added setVolume, setPosition, and play now resumes.
setVolume(vol) sets a volume of the audio, 0 is off, 1 is normal
loudness, this is a double.  setPosition(percent) sets the current time
of the song to a percentage, 0 is beginning, 1 is ending.  Play() now
resumes from last stopped part.  We need to add a resume button to
differentiate the two.