Added Volume Slider and Position Slider
Added Volume Slider and Position Slider to the control box, also fixed
the issue on the setPosition function where it set the position based on
the offset of the current elapsed time instead of the initial time of 0.
Created the resume button to call the Resume function. So we can modify
the play function to start from 0.