Added Time callbacks, and current position.
positionCallback returns the current percentage, and it is also
automatically called every second, this can be used to move the position
slider.  When the song ends naturally, the position is moved to 0, and
the songEnded() function is called.  This will be used later for
changing to the next song or displaying information.