Added Song Queue
Added a song queue that shows a playlist of songs that are loaded. It
will load the next song when one ends, you can play any song by clicking
on it in the queue, and the current song playing is highlighted green.
The variable that holds the song is called songQueue and the variable
currentSong is the number of what song is currently playing in the
queue.

The queue is located in the advance control panel, it will scroll the
number of songs displayed get past a certain height.