Added Audio plans to Drawings Folder
This is how the overall hierarchy should look like by the end of this
sprint (2).  Sources are made from buffers.  Sources connect to a gain
node, used to crossfade between songs.  Crossfade nodes will merge to an
array of BiQuadFilter nodes.  Each biquad node will have a different
effect on the song.  The first filter in the array should be connected
with the cross fade gains, while the last filter node in the array
should connect to both the analyzer node and the volume node.  The
analyzer connects to a javascript node, used to interact with Three.JS,
which is then connected to the final destination.  The volume node
connects directly to the destination, and avoids the analyzer.