Implemented the "songs" object, start using it.
the songs object has a "now", "next" and "last".  Each of those has a
source, buffer, gain object, and index.  The idea is that this is a hash
for the previous songs, the current song, and the next song.  This
allows us to mix and crossfade songs for smooth transitions.

Additionally, the method "requestSong" has been changed.  It will now
take parameters of (url, songObject), and will load a url into the
selected songObject.  The songObject is either songs.now, songs.next, or
songs.last.  This doesn't change anything yet, but will be vital to the
next change.