Next and Previous buttons now work.
Currently, next and last buffers will be pre-loaded automatically.  Once
they are loaded, you can use the next/previous buttons.  If the songs
aren't loaded, you must wait for them to load before clicking on the
buttons.  This might change in the future but this is how it will be for
now.

Additionally, theres a minor bug where last/next won't preload with
songs you add to the queue later.  I'll try to get this done tomorrow.