Created a new demo for thumbnails
We can create a mini-rendering of our scene by appending the domElement
of the renderer to a div.  Additionally, we must change the camera
fov/aspect ratio to be equal to this thumbnail's width / height.

This will probably make it into sprint three if we have more than one
scene!