Updated UI
Updated the UI to look better, all buttons work the same, no ID's have
changed (except loadingDialog changed to loadingIcon). Added functions:
showQueue and hideQueue which causes the panel on the left to open to
show the songQueue. Also added the function appendSong which will add a
song to the queue visually