Created a shader for our frequency visualizer
The js\shaders director will be where we store shaders, please use the
naming convention of "shader-name-fragment" or "shader-name-vertex" for
vertex vs fragment shaders.

We can load the shaders using the new function: loadShader(filepath).
This is simply an ajax request to load a file as text, which is what a
shader takes as input.