Added time uniform to fragment shader
The time uniform allows the color of the material to change over time.
We can create a loop of colors based on a sin function of time.  The
amplitude of this can be changed, as well as the speed at which it
changes.  This technique will be vital to creating beautiful, dynamic
colors in all scenes.