Added Sound Cloud Form and Advance Controls menu
Added an Advance Controls Menu to the right hand side. Slides out like
the songQueue, unfortunately a limitation in the plugin results in
Advance Controls and Song Queue to not be open at the same time (so not
a bug). Added sound cloud forms for username and password as well as a
url box for the url. They do not look pretty yet, but at least they work
(which was harder than it should of been...).