Scene select is now working, need divs to place
Scene selection is handled in the scene-helper.js file.  This file has
methods to change scenes, and append them to certain divs.

Scenes must now follow a certain structure to be properly added to the
project.  See the scene-helper file for more info, (just try to copy
scene1.js)

Lucas:  We need a place to store the thumbnails.