Updated scene1 to not change on pause state
Now the scene won't disform when you pause it.  Additionally made the
scene boxes have a set size.  We'll have to change how scene select
works, I think we should create a relative positioned div on the top of
the screen, so all scenes are on the screen at once.  This should
probably fix the weird bugs with appending scenes to the advanced
controls menu.