Added Function Modifier Modules
Added the mod select and functional modifier modules with the proper
sliders that they take. The mods are identified by the id 'mod-#' with
'#" being the child number in the dom. The type can be accessed by
reading the title or by using jquery to access the type data attr
$("#mod-1").data('type').